Implemented code from my testing area into task1_6
